requirements_graphrag.txt
openai>=1.0.0
neo4j>=5.13.0
langchain>=0.1.0
llama-index>=0.10.0
langgraph>=0.0.10
streamlit>=1.18.0
python-dotenv